+++ location_tracker/URL.swift
@@ -14,13 +14,17 @@ class URL{
         })
     }
 
-    func downloadImage(url: NSURL, handler: ((image: UIImage, NSError!) -> Void))
+    func downloadImage(url: NSURL, handler: ((image: UIImage?, NSError!) -> Void))
     {
         var imageRequest: NSURLRequest = NSURLRequest(URL: url)
         NSURLConnection.sendAsynchronousRequest(imageRequest,
             queue: NSOperationQueue.mainQueue(),
             completionHandler:{response, data, error in
-                handler(image: UIImage(data: data)!, error)
+                if error != nil || data == nil {
+                    handler(image: nil, error)
+                    return
+                }
+                handler(image: UIImage(data: data), error)
         })
     }